==> notebooks/1_data_summary.ipynb <==
# Data Summary
#### John R. Starr; jrs294@pitt.edu

The data is split into two folders/files, en/TEP.xml and fa/TEP.xml. Let's import what we need to properly search through this data:

import pandas as pd
import numpy as np
import nltk
import xml.etree.ElementTree as ET
from lxml import etree

Making sure that we're in the correct directory:

import os
os.getcwd()

First, we need to build a parser works for xml. I found documentation on etree.XMLParser() [here](https://lxml.de/api/lxml.etree.XMLParser-class.html). After some preliminary efforts in building trees, I found that some of my data has corrupted characters (or at least something along those lines). An example of one of these encoding errors can be found in the following sentence: <s id="86377">simple caf oronary . freak show choked to death .</s> When opened in Notepad++, the space between "caf" and "oronary" is the abbreviation NUL highlighted in black. Other problems occur later in the dataset.

In order to get my parser to work, I've hand-modified the dataset. Any modifications that I made can be found in the data_modifications.txt file [here](https://github.com/Data-Science-for-Linguists-2019/Scrambling-in-English-to-Persian-Subtitles/blob/master/data_modifications.txt). The name of this edited file is 'TEP_mod.xml' and will be used for the remainder of this project.

# Creating a new parser
parser_full = etree.XMLParser(recover = True)
tree_eng = ET.parse('Private/TEP/raw/en/TEP_mod.xml', parser = parser_full)

Now, let's build the root and see how our data is structured. I've looked through the XML file and noticed that there is a bit of a heading that looks like this (I have added spaces between the greater/less than symbols so that it remains visible in this file):

< ?xml version="1.0" encoding="utf-8"? >
< letsmt version="1.0" >
< head >< /head >
    
After that, we have the body character, followed by a sentence. So, we'll use .findall() to start where we want it to start, and hopefully we'll be able to get an idea of what our data looks like:

root_eng = tree_eng.getroot()
root_eng.items()
for item in root_eng.findall('./body/s')[:5]:
    print(item.text)
    #print(dir(item))
    print(item.values())
    print(item.items())

It looks like our data uses the ID number to mark what text comes after it. This works well, as we'll be able to match up the keys between the two files to combine them! Let's create a test dictionary in which the key is the item and the text is the value, replacing all extraneous information:

eng_lines_test = {}
for item in root_eng.findall('./body/s')[:5]:
     eng_lines_test[int(str(item.values()).replace(',', '').replace("['", '').replace("']", ''))] = str(item.text.replace(',', '').replace(' .', ''))

eng_lines_test.keys()

eng_lines_test.values()

Awesome! Let's do the same for the Farsi text as well!

tree_far = ET.parse('Private/TEP/raw/fa/TEP.xml', parser = parser_full)

far_lines_test = {}
root_far = tree_far.getroot()
for item in root_far.findall('./body/s')[:5]:
    far_lines_test[int(str(item.values()).replace(',', '').replace("['", '').replace("']", ''))] = str(item.text.replace(',', '').replace(' .', ''))

far_lines_test.keys()

far_lines_test.values()

All in the clear! Now, let's combine the two into a single DataFrame object!

test_DF = pd.Series(eng_lines_test).to_frame('eng').join(pd.Series(far_lines_test).to_frame('far'), how='outer')
test_DF

It worked! Yay! Let's apply this methodology for both of the files in full, rather than the little pieces we've been testing:

eng_lines = {}
for item in root_eng.findall('./body/s'):
     eng_lines[int(str(item.values()).replace(',', '').replace("['", '').replace("']", ''))] = str(item.text.replace(',', '').replace(' .', ''))
print(len(eng_lines))

far_lines = {}
for item in root_far.findall('./body/s'):
     far_lines[int(str(item.values()).replace(',', '').replace("['", '').replace("']", ''))] = str(item.text.replace(',', '').replace(' .', ''))
print(len(far_lines))

Cool! We have the same numbers. Let's see what the DF would look like, and then add some more information that might be useful to use in the future.

full_df = pd.Series(eng_lines).to_frame('Eng').join(pd.Series(far_lines).to_frame('Far'), how='outer')
full_df.index.name = 'ID'
full_df

full_df.describe()

Some interesting things to point out:
- It appears that the subtitles include non-spoken acts of communication, such as "raspy breathing" or "music playing". I'm not entirely sure how to remove this data, as it is not marked in any particular way.
- Also, poor William in the beginning! It seems that he's having a tough time...
- "Yes" and it's Persian translation "بله" are the most common words, but they do not occur at the same frequency... This may be because it is common for Persian to repeat "بله" when speaking casually.

Back to the data. Let's create three more columns for both langauges: token, token count (or length), and type. This will help us navigate the data when performing analysis:

# Constructing Token columns
full_df['Eng_Tok'] = full_df['Eng'].apply(nltk.word_tokenize)
full_df['Far_Tok'] = full_df['Far'].apply(nltk.word_tokenize)

# Constructing Len columns
full_df['Eng_Len'] = full_df['Eng_Tok'].apply(len)
full_df['Far_Len'] = full_df['Far_Tok'].apply(len)

# Constructing Type columns
full_df['Eng_Types'] = full_df['Eng_Tok'].apply(set)
full_df['Far_Types'] = full_df['Far_Tok'].apply(set)

Seeing our resulting DF:

full_df.head()

How many one-word lines are there? If a significant portion of my data consists of one-word lines, then it will be pretty challenging to get results analyzing the syntax of the subtitles. I'm going to predict that there will be more Persian one-word lines, as you don't need to include the subject in Persian and can simply utter a verb. 

# Seeing how many one-word lines there are in English, Farsi, and both.
eng_1word = [x for x in full_df['Eng_Len'] if x == 1]
print(len(eng_1word))

far_1word = [x for x in full_df['Far_Len'] if x == 1]
print(len(far_1word))

both_1word = [x for x in full_df if len(full_df['Eng_Len']) == 1 if len(full_df['Far_Len']) == 1]
print(len(far_1word))

# Overall length of file:
len(full_df)

Interesting! It looks like my hypothesis was wrong. Every Persian line that is one word is also one word in English. However, reverse is not true. This is something I will investigate further in my data analysis. 

These sentences do not take up a significant portion of my data, but I am considering removing them because they do not help me with my analysis. I will be meeting with one of the instructors to get a second opinion on this matter.

Let's pickle the data as a whole and move it to my private folder, pickle a small portion of the data, get a little more information on the data, and then move on to POS tagging and shallow parsing!

full_df.to_pickle('full_df.pkl') # This will be put in my private folder. 

# Seeing general information about the DF
full_df.info()

Well, we don't have any null values! This is good (and expected). What's the average sentence length for each language?

full_df.Eng_Len.value_counts()

full_df.Far_Len.value_counts()

# Average English sentence length
eng_len_tot = []
for item in full_df['Eng_Len']:
    eng_len_tot.append(item)
eng_len_avg = (sum(eng_len_tot))/len(full_df)
print(eng_len_avg)

# Average Farsi sentence length
far_len_tot = []
for item in full_df['Far_Len']:
    far_len_tot.append(item)
far_len_avg = (sum(far_len_tot))/len(full_df)
print(far_len_avg)

It seems that the sentences are pretty comparable in average length. This conclusion can be inferred from looking at the overall value counts for each of the languages. 

I think that's it for data exploration and summarization. Let's move to tagging and parsing the data!
