==> notebooks/2_tagging_chunking.ipynb <==
# POS Tagger and Shallow Parsing
#### John R. Starr; jrs294@pitt.edu
In this file, I will apply my POS tagger to my data. Following this, I will use a shallow parser to determine the grammatical structure for both English and Persian sentences. 

The reason that this is a separate file is because my laptop runs on a Windows environment. After a lot of struggle and problem-solving, Dan suggested I try Colab. Thankfully, Colab works! So, I ran the code that wasn't working on my laptop, uploaded the necessary files, and got everything to work! As a result, I was able to POS-tag and shallow parse my data. 

I am using the Hazm library to manipulate my Persian text; information about this corpus can be found [here]('https://github.com/sobhe/hazm').

!pip install hazm

Following here, I will be using the general pipeline that the creators outline in their repository:

### POS-tagging and Chunking Persian Text

# Importing the necessary modules
from hazm import *

# Building our tagger
tagger = POSTagger(model='postagger.model')
tagger.tag(word_tokenize('ما بسیار کتاب می‌خوانیم'))

# Building our tagger
chunker = Chunker(model='chunker.model')

# Test file for the chunker
tagged = tagger.tag(word_tokenize('کتاب خواندن را دوست داریم'))
tree2brackets(chunker.parse(tagged))

It works! Sweet. Let's read in our pickled DF from our data_summary file and tag our Persian sentences.

import pandas as pd

# Reading in the file
full_df = pd.read_pickle('full_df.pkl')
full_df.head()

# Creating our POS column for Farsi
full_df['Far_POS'] = full_df['Far_Tok'].apply(tagger.tag)

# Creating our Chunks column based on the POS column
full_df['Far_Chunks'] = full_df['Far_POS'].apply(chunker.parse).apply(tree2brackets)

full_df.head()

Sweet! It looks like that we've got all the Persian tagging and chunking in order. Now, we need to do the same for the English words, using nltk's modules. These may not be the best modules, but they are the only ones that I can seem to get to work... I intend on scheduling at least a few OH appointments so that I can get a better POS-tagger and chunker for English up and running.

Anyway, let's import nltk and download all the necessary files.

### POS-tagging and Chunking English Text

import nltk
nltk.download('averaged_perceptron_tagger')
nltk.download('maxent_ne_chunker')
nltk.download('words')

# Checking the default POS tagger in nltk:
nltk.pos_tag('maybe the wind is causing it. i hate it'.split())

# Seeing how chunk.ne_chunk works...
tagged_sentence = nltk.pos_tag('maybe the wind is causing it. i hate it'.split()) #pos_tagging the sentence
result = nltk.chunk.ne_chunk(tagged_sentence) #chunking the sentence
print(result)

So this didn't really work... what about this parser?

# Second Parser off of NLTK
from nltk.parse import ParserI
from nltk.chunk.api import ChunkParserI

p = ChunkParserI()
p.parse(tagged_sentence)

Hmm... this didn't work either... I need to get these to work (expect a few trips to OHs!)

For now, since I have to hand something in, let's just add the Eng_POS column and pickle what we have to a new pkl file that we can use in our analysis.

full_df['Eng_POS'] = full_df['Eng_Tok'].apply(nltk.pos_tag)

full_df.head()

Let's pickle it:

full_df.to_pickle('full_df_MOD.pkl')

That's it for now! Time for some analysis.

#### NOTE
I am very frustrated with myself and my project and will be coming in to OH to figure things out and recenter.
