==> notebooks/3_chunking_english.ipynb <==
# Chunking English Files
#### John R. Starr; jrs294@pitt.edu
Since I had trouble handling my various parsers and chunkers for the previous progress report (due to some technical limitations of my laptop), I will load the pickle file that I had previously created, rather than add on to the previous JPYNB found [here](https://github.com/Data-Science-for-Linguists-2019/Scrambling-in-English-to-Persian-Subtitles/blob/master/tagging_chunking.ipynb).

So, let's chunk the English files!!

from nltk.chunk.regexp import *
import nltk
import pandas as pd
import numpy as np

I've been messing around with various parsers for English, but this one seems to run the best:

parser = RegexpParser('''
    NP: {<DT>? <JJ>* <NN>* <P>*} # NP
    P: {<IN>}           # Preposition
    V: {<V.*>}          # Verb
    PP: {<P> <NP>}      # PP -> P NP
    VP: {<V> <NP|PP>*}  # VP -> V (NP|PP)*
    ''')

Let's load the pickle file...!

full_df = pd.read_pickle('full_df_MOD.pkl')
full_df.head(10)

Let's see an example parse:

print(parser.parse(full_df.Eng_POS.iloc[8]))

full_df['Eng_Chunks'] = full_df['Eng_POS'].apply(parser.parse)

full_df.head(10)

full_df.to_pickle('tagged_chunked_df.pkl')
